2nd version

+++ ProjetWebTroc/src/EventBundle/Controller/EventController.php
@@ -10,7 +10,10 @@ use EventBundle\Form\EventUpdateType;
 use EventBundle\Repository\ParticipantRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
+use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\Serializer\Normalizer\ObjectNormalizer;
+use Symfony\Component\Serializer\Serializer;
 use TrocBundle\Entity\Event;
 use EventBundle\Entity\Participant;
 
@@ -21,42 +24,51 @@ class EventController extends Controller
         $error = null;
         if ($this->getUser() != null) {
             $events = $this->getDoctrine()->getRepository(Event::class)->findAll();
+            //$events = $this->getDoctrine()->getRepository(Participant::class)->CountEvents();
+            $current = new \DateTime ("now");
             $paginator = $this->get('knp_paginator');
             $pagination = $paginator->paginate(
                 $events, /* query NOT result */
                 $request->query->getInt('page', 1)/*page number*/,
                 5 /*limit per page*/);
 
-            if(isset($_GET['join']))
+            if (isset($_GET['join']))
             {
                 $idEvent = $_GET['id'];
                 $reserve = new Participant();
 
                 $verify = $this->getDoctrine()
-                                ->getManager()
-                                ->getRepository(Participant::class)
-                                ->verifyJoin($this->getUser()->getId(), $idEvent);
-                if($verify == null)
+                    ->getManager()
+                    ->getRepository(Participant::class)
+                    ->verifyJoin($this->getUser()->getId(), $idEvent);
+                if ($verify == null)
                 {
                     //die('does not exist');
                     $reserve->setIdUser($this->getUser()->getId());
                     //$reserve->getIdUser()
-                    $reserve->setIdEvent($idEvent);
+                    $event = $this->getDoctrine()->getRepository(Event::class)->find($idEvent);
+
+                    $reserve->setIdEvent($event);
 
                     $em = $this->getDoctrine()->getManager();
                     $em->persist($reserve);
                     $em->flush();
+                    $em = $this->getDoctrine()->getManager();
+                    $event = $em->getRepository(Event::class)->find($idEvent);
+                    $event->setNbr($event->getNbr()-1);
 
+                    $em->flush();
                     $error = 1;
 
-                }
-                else{
-                    $error =2;
+                } else
+                    {
+                    $error = 2;
                 }
             }
             return $this->render('@Event/Event/list.html.twig', array(
                 'events' => $pagination,
                 'error' => $error,
+                'current' => $current,
 
             ));
 
@@ -67,40 +79,55 @@ class EventController extends Controller
 
     public function createEventAction(Request $request)
     {
-        $event = new Event();
-        $userID = $this->getUser()->getId();
-        $form = $this->createForm(EventType::class, $event);
-        $form = $form->handleRequest($request);
-        if ($form->isValid()) {
-            /**
-             * @var UploadedFile $file
-             */
-            $file=$event->getImage();
-            $fileName = md5(uniqid()).'.'.$file->guessExtension();
-            $file->move($this->getParameter('image_directory'),$fileName);
-            $event->setImage($fileName);
-            $em = $this->getDoctrine()->getManager();
-            $event->setCreatedby($userID);
-            $em->persist($event);
-            $em->flush();
+        if ($this->getUser() != null) {
+            $event = new Event();
+            $userID = $this->getUser()->getId();
+            $form = $this->createForm(EventType::class, $event);
+            $form = $form->handleRequest($request);
+            if ($form->isValid()) {
+                /**
+                 * @var UploadedFile $file
+                 */
+                $file = $event->getImage();
+                $fileName = md5(uniqid()) . '.' . $file->guessExtension();
+                $file->move($this->getParameter('image_directory'), $fileName);
+                $event->setImage($fileName);
+                $em = $this->getDoctrine()->getManager();
+                $event->setCreatedby($userID);
+                $event->setNbr($event->getMax());
+                $em->persist($event);
+                $em->flush();
+            }
+            return $this->render('@Event/Event/CreateEvent.html.twig', array('form' => $form->createView()));
+        } else {
+            return $this->redirectToRoute('fos_user_security_login');
         }
-        return $this->render('@Event/Event/CreateEvent.html.twig', array('form' => $form->createView()));
 
     }
 
     public function deleteEventAction($id)
     {
-        $em=$this->getDoctrine()->getManager();
-        $event=$em->getRepository(Event::class)->find($id);
-        $em->remove($event);
-        $em->flush();
+        if ($this->getUser() != null) {
+            $em = $this->getDoctrine()->getManager();
+            $event = $em->getRepository(Event::class)->find($id);
+            $em->remove($event);
+            $em->flush();
 
-        return $this->redirectToRoute('event_listEvent');
+            return $this->redirectToRoute('event_listEvent');
+        } else {
+            return $this->redirectToRoute('fos_user_security_login');
+        }
+
     }
 
-    public function choiceHomeAction ()
+    public function choiceHomeAction()
     {
-        return $this->render('@Event/Event/choiceHome.html.twig');
+        if ($this->getUser() != null) {
+            return $this->render('@Event/Event/choiceHome.html.twig');
+        } else {
+            return $this->redirectToRoute('fos_user_security_login');
+        }
+
     }
 
 
@@ -109,8 +136,8 @@ class EventController extends Controller
 
         $user = $this->getUser();
         $idUser = $user->getId();
-        $idEvent =  $event->getId();
-        die('RR: '.$idEvent );
+        $idEvent = $event->getId();
+
         $partic = new Participant();
         $partic->setIdEvent($idEvent);
         $partic->setIdUser($idUser);
@@ -119,16 +146,15 @@ class EventController extends Controller
             $form->persist($partic);
             $form->flush();
 
-        }
-        else
-        {
-            return $this->redirectToRoute('troc_signin');
+        } else {
+            return $this->redirectToRoute('fos_user_security_login');
         }
     }
 
     public function myEventAction(Request $request)
     {
-        if ($this->getUser() != null) {
+        if ($this->getUser() != null)
+        {
             $id = $this->getUser()->getId();
             $events = $this->getDoctrine()->getRepository(Event::class)->findBygetMyEvents($id);
 //            die('sd: '.$events);
@@ -140,39 +166,70 @@ class EventController extends Controller
 
             return $this->render('@Event/Event/myEvents.html.twig', array(
                 'events' => $pagination));
-        }
-        else
-        {
+        } else {
             return $this->redirectToRoute('fos_user_security_login');
         }
     }
 
-    public function goingToAction ()
+    public function goingToAction()
     {
-
-        $user = $this->getUser()->getId();
-        $events= $this->getDoctrine()->getRepository(Event::class)->getMyListEvents($user);
-        return $this->render('@Event/Event/goingTo.html.twig',array('events'=>$events));
-    }
-
-    public function updateEventAction($id,Request $request)
-    {
-        $em=$this->getDoctrine()->getManager();
-        $event=$em->getRepository(Event::class)->find($id);
-        $form=$this->createForm(EventUpdateType::class,$event);
-        $form=$form->handleRequest($request);
-        if($form->isValid())
+        if ($this->getUser() != null)
         {
-
-            $em->flush();
+            $user = $this->getUser()->getId();
+            $events = $this->getDoctrine()->getRepository(Event::class)->getMyListEvents($user);
+            return $this->render('@Event/Event/goingTo.html.twig', array('events' => $events));
+        } else {
+            return $this->redirectToRoute('fos_user_security_login');
         }
-        return $this->render('@Event/Event/update.html.twig', array('form'=>$form->createView()
-        // ...
-        ));
     }
 
-    public function searchEventAction()
+    public function updateEventAction($id, Request $request)
     {
+        if ($this->getUser() != null)
+        {
+            $em = $this->getDoctrine()->getManager();
+            $event = $em->getRepository(Event::class)->find($id);
+            $form = $this->createForm(EventUpdateType::class, $event);
+            $form = $form->handleRequest($request);
+            if ($form->isValid()) {
+
+                $em->flush();
+            }
+            return $this->render('@Event/Event/update.html.twig', array('form' => $form->createView()));
+        } else {
+            return $this->redirectToRoute('fos_user_security_login');
+        }
+    }
+    public function updatePartAction($id, Request $request)
+    {
+        if ($this->getUser() != null)
+        {
+            $em = $this->getDoctrine()->getManager();
+            $event = $em->getRepository(Event::class)->find($id);
+            $event->setNbr($event->getNbr()-1);
+
+                $em->flush();
+            }
+
+    }
+
+    public function searchEventAction(Request $request)
+    {
+        if($request->isXmlHttpRequest())
+        {
+            $name=$request->get('srch');
+
+            $events=$this->getDoctrine()->getRepository(Event::class)->findBy(array('titre'=>$name));
+
+            //SERIALISATION
+            //initialisation
+            $se=new Serializer(array(new ObjectNormalizer()));
+            //normalize de la liste
+            $data=$se->normalize($events);
+            return new JsonResponse($data);
+
+        }
+
         return $this->render('@Event/Event/searchEvent.html.twig');
     }
 }
